Revert "fix ResultRepository"
This reverts commit 8c8a3fd7d3693138252347ccacc3ff96a59e869b.

+++ backend/src/main/java/com/example/TKPM_Project/repository/ResultRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResultRepository extends JpaRepository<Result, Long> {
-    // Có thể thêm các phương thức truy vấn tùy chỉnh nếu cần
+public interface ResultRepository extends JpaRepository<Result, Integer> {
+
 }
